feat: Add Eyed3TagManager for reading ID3 tags
The Eyed3TagManager class is added to the project, which implements the TagManagerInterface interface. It provides a method, read_id3_tags(), to read ID3 tags from MP3 files using the Eyed3 library. This allows retrieving the title, artist, album, and track number from MP3 files.

This commit also includes the necessary changes to the requirements files and the addition of the TagCollection class.

Note: This commit message does not include any meta information like issue references, tags, or author names.